Migrate and split up test projects
