+++ tests/JWT.Tests.Common/Customer.cs
@@ -1,0 +1,9 @@
+﻿namespace JWT.Tests.Common
+{
+    public class Customer
+    {
+        public string FirstName { get; set; }
+
+        public int Age { get; set; }
+    }
+}
