+++ tests/JWT.Tests.Common/TestData.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections.Generic;
+
+namespace JWT.Tests.Common
+{
+    public static class TestData
+    {
+        public static readonly Customer Customer = new Customer { FirstName = "Bob", Age = 37 };
+
+        public const string Token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJGaXJzdE5hbWUiOiJCb2IiLCJBZ2UiOjM3fQ.cr0xw8c_HKzhFBMQrseSPGoJ0NPlRp_3BKzP96jwBdY";
+        public const string MalformedToken = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9eyJGaXJzdE5hbWUiOiJCb2IiLCJBZ2UiOjM3fQ.cr0xw8c_HKzhFBMQrseSPGoJ0NPlRp_3BKzP96jwBdY";
+        public const string ExtraHeadersToken = "eyJmb28iOiJiYXIiLCJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJGaXJzdE5hbWUiOiJCb2IiLCJBZ2UiOjM3fQ.slrbXF9VSrlX7LKsV-Umb_zEzWLxQjCfUOjNTbvyr1g";
+
+        public static readonly IDictionary<string, object> DictionaryPayload = new Dictionary<string, object>
+        {
+            { "FirstName", "Bob" },
+            { "Age", 37 }
+        };
+    }
+}
