--- tests/JWT.Tests/Internal/Customer.cs
@@ -1,9 +1,0 @@
-﻿namespace JWT.Tests
-{
-    internal class Customer
-    {
-        public string FirstName { get; set; }
-
-        public int Age { get; set; }
-    }
-}
