--- tests/JWT.Tests/packages.config
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentAssertions" version="4.19.2" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="10.0.2" targetFramework="net461" />
-  <package id="ServiceStack.Text" version="4.5.8" targetFramework="net461" />
-  <package id="xunit" version="2.2.0" targetFramework="net461" />
-  <package id="xunit.abstractions" version="2.0.1" targetFramework="net461" />
-  <package id="xunit.assert" version="2.2.0" targetFramework="net461" />
-  <package id="xunit.core" version="2.2.0" targetFramework="net461" />
-  <package id="xunit.extensibility.core" version="2.2.0" targetFramework="net461" />
-  <package id="xunit.extensibility.execution" version="2.2.0" targetFramework="net461" />
-  <package id="xunit.runner.visualstudio" version="2.2.0" targetFramework="net461" developmentDependency="true" />
-</packages>
